Moved make_bone_indices_consistent to vertex_fusion for testing purposes
